ss: support concurrent execution
When using goroutines, the behavior of CreateComDetailsFromNode resulted
in the deletion of the debug namespace, causing failure.

Here, we relocate the creation and deletion of the debug namespace
outside of the function to support running this function in parallel threads.
